requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiohttp>=3.9.0
sqlalchemy>=2.0.25
aiomysql>=0.2.0
wavelink>=3.4.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-socketio>=5.11.0
python-multipart>=0.0.6
jinja2>=3.1.3
authlib>=1.3.0
itsdangerous>=2.1.2
loguru>=0.7.2
apscheduler>=3.10.4
humanize>=4.9.0
python-dateutil>=2.8.2